0 task

+++ 0x12-singly_linked_lists/lists.h
@@ -8,7 +8,7 @@
 struct list
 {
 char *str;
-long unsigned int len;
+unsigned int len;
 struct list *next;
 };
 typedef struct list list_t;
